Persist the Preview URL for future jobs
Having the Preview URL available could be useful to run other
checks on the Preview, for example an a11y audit.

Here's how to use the persisted file that contains the URL into
another job:

  steps:
    - attach_workspace:
        # Must be absolute path or relative path from working_directory
        at: /tmp/workspace

    - run: |
        if [[ `cat /tmp/workspace/preview-url` == "http://preview-url.com" ]]; then
          echo "It worked!";
        else
          echo "Nope!"; exit 1
        fi

Of course the job that wants to use the Preview URL needs to require
this orb to run into a previous step before it to start (using require).

Here's more info:

https://circleci.com/docs/2.0/workflows/#using-workspaces-to-share-data-among-jobs