Remove the $DIRECTORY folder immediately after the creation of the image
After the creation of the docker image, the $DIRECTORY folder is not
required anymore. Instead of removing the old directories during the
`Cleanup Old Containers` step, it is better to remove them at the end of
the `Build Docker Image On Docker Server` step, freeing up space on the
server.