Execute the `docker rm` only if the `old_containers` file is not empty
When there is nothing to delete and the `old_containers` file is empty,
the `docker rm` command returns the following error:

"docker rm" requires at least 1 argument.

and this causes the CI to fail.

This commit adds a check that executes the `docker rm` command only if
the `old_container` file is not empty.